Changes to comments and html to allow comments to be stored correctly

+++ src/main/java/com/makersacademy/acebook/repository/CommentRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface CommentRepository extends CrudRepository<Comment, Integer> {
+public interface CommentRepository extends CrudRepository<Comment, Long> {
     public Optional<Comment> findCommentByUsername(String username);
-    public Optional<Comment> findCommentByPostID(int postID);
+    public Optional<Comment> findCommentByPostID(Integer postID);
 
 }
